% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancRscore.R
\docType{data}
\name{cancRscore}
\alias{cancRscore}
\title{Cancer-related score derived from an Elastic Net binomial regression model, based on a dataset aggregated from multiple cancer-focused databases, including COSMIC, PubMed, ClinicalTrials, TCGA, GTEx, and DepMap}
\format{
A data frame with columns:
\describe{
\item{pubmed_gene_symbol}{HUGO gene symbols extracted from pubmed dataset.}
\item{cosmic_cgc_status}{Boolean for genes curated as cancer-related or not cancer-related.}
\item{entrez_id}{Gene Entrez ID.}
\item{s1}{Probability score extracted from the Elastic Net binomial regression model.}
\item{other columns}{Check datasets extraction and treatment pipeline \url{https://github.com/Karksus/uspesalq_project/tree/master}.}
}
}
\source{
\url{https://github.com/Karksus/uspesalq_project/tree/master}
}
\usage{
cancRscore
}
\description{
Cancer-related score derived from an Elastic Net binomial regression model, based on a dataset aggregated from multiple cancer-focused databases, including COSMIC, PubMed, ClinicalTrials, TCGA, GTEx, and DepMap
}
\examples{
cancRscore

}
\keyword{datasets}
