% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicaltrials_genes.R
\docType{data}
\name{clinicaltrials_genes}
\alias{clinicaltrials_genes}
\title{Clinicaltrials gene citation count}
\format{
A data frame with columns:
\describe{
\item{clinicaltrials_gene_symbol}{HUGO gene symbols.}
\item{clinicaltrials_StudyCount}{Gene citation count at clinicaltrials.gov.}
\item{entrez_id}{Gene Entrez ID.}
}
}
\source{
\url{https://clinicaltrials.gov/}, extracted Jun 25, 2024
}
\usage{
clinicaltrials_genes
}
\description{
Clinicaltrials gene citation count
}
\examples{
clinicaltrials_genes

}
\keyword{datasets}
