% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depmap_crispr.R
\docType{data}
\name{depmap_crispr}
\alias{depmap_crispr}
\title{CRISPR 22Q1 DepMap depedency scores.}
\format{
A data frame with columns:
\describe{
\item{gene_name}{HUGO gene symbols..}
\item{entrez_id}{Gene Entrez ID.}
\item{other columns}{Cancer cell-lines.}
}
}
\source{
\url{https://oncodb.org/index.html}, downloaded Oct 12, 2024
}
\usage{
depmap_crispr
}
\description{
CRISPR 22Q1 DepMap depedency scores.
}
\examples{
depmap_crispr

}
\keyword{datasets}
