% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depmap_rnai.R
\docType{data}
\name{depmap_rnai}
\alias{depmap_rnai}
\title{RNAi 19Q3 DepMap depedency scores.}
\format{
A data frame with columns:
\describe{
\item{gene_name}{HUGO gene symbols..}
\item{entrez_id}{Gene Entrez ID.}
\item{other columns}{Cancer cell-lines.}
}
}
\source{
\url{https://oncodb.org/index.html}, downloaded Oct 12, 2024
}
\usage{
depmap_rnai
}
\description{
RNAi 19Q3 DepMap depedency scores.
}
\examples{
depmap_rnai

}
\keyword{datasets}
