% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{nocgc_genes}
\alias{nocgc_genes}
\title{Get top no-cgc genes}
\usage{
nocgc_genes(top = 10, threshold = 0.5)
}
\arguments{
\item{top}{A number.}

\item{threshold}{A number.}
}
\value{
A data frame.
}
\description{
Get top no-cgc genes
}
\examples{
nocgc_genes(10, 0.5)
}
