% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_cagr.R
\docType{data}
\name{pubmed_cagr}
\alias{pubmed_cagr}
\title{Compound Annual Growth Rate (CAGR) score of genes citation growth 2013-2023.}
\format{
A data frame with columns:
\describe{
\item{pubmed_gene_symbol}{HUGO gene symbols..}
\item{entrez_id}{Gene Entrez ID.}
\item{pubmed_CAGR}{CAGR score.}
}
}
\source{
\url{https://oncodb.org/index.html}, downloaded Oct 12, 2024
}
\usage{
pubmed_cagr
}
\description{
Compound Annual Growth Rate (CAGR) score of genes citation growth 2013-2023.
}
\examples{
pubmed_cagr

}
\keyword{datasets}
