% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_gtex_geexp.R
\docType{data}
\name{tcga_gtex_geexp}
\alias{tcga_gtex_geexp}
\title{Tumor and normal tissue gene expression extracted from TCGA and GTEX dataset.}
\format{
A data frame with columns:
\describe{
\item{cancer_type}{TCGA cancer types.}
\item{gene_expression_status}{Level of tumor/normal tissue gene expression .}
\item{entrez_id}{Gene Entrez ID.}
}
}
\source{
\url{https://oncodb.org/index.html}, downloaded Oct 12, 2024
}
\usage{
tcga_gtex_geexp
}
\description{
Tumor and normal tissue gene expression extracted from TCGA and GTEX dataset.
}
\examples{
tcga_gtex_geexp

}
\keyword{datasets}
