% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_gtex_meth.R
\docType{data}
\name{tcga_gtex_meth}
\alias{tcga_gtex_meth}
\title{Tumor and normal tissue methylation levels extracted from TCGA and GTEX dataset.}
\format{
A data frame with columns:
\describe{
\item{cancer_type}{TCGA cancer types.}
\item{methylation_status}{Level of tumor/normal methylation.}
\item{entrez_id}{Gene Entrez ID.}
}
}
\source{
\url{https://oncodb.org/index.html}, downloaded Oct 12, 2024
}
\usage{
tcga_gtex_meth
}
\description{
Tumor and normal tissue methylation levels extracted from TCGA and GTEX dataset.
}
\examples{
tcga_gtex_meth

}
\keyword{datasets}
